+++ src/main/java/net/danygames2014/unitweaks/util/ModOptions.java
@@ -15,13 +15,16 @@ public class ModOptions {
     public static Option brightnessOption;
     public static Option guiScaleOption;
 
+    // Spawn Protection
+    public static int spawnProtection = 16;
+    
     // Front View Third Person
     public static boolean frontView = false;
     
     // GUI Scale
     public static float guiScale = 0.0F;
     public static float realGuiScale = 0.0F;
-
+    
     public static int getGuiScale() {
         return (int) Math.floor(realGuiScale * 8);
     }